Add files via upload
Independent Study of MSIB Batch 5 on Startup Campus  
On this occasion I managed to join the Data Science Track. For approximately 6 months we focus on learning from the basic introduction of "Intro to Data Science" to "Machine Learning". Learning with mentors and practitioners is doing by online, and at the end of the semester there is a final project as a form of learning output. 
Here is the description of the Final Project Team Zeus